backend/requirements.txt
langgraph==0.2.35
requests==2.32.3
tenacity==9.0.0
psycopg2-binary==2.9.9
ortools==9.10.4067
pydantic==2.8.2
python-dateutil==2.9.0.post0
